+++ core/src/main/java/by/st/deliver/core/servicesImpl/RestaurantServiceImpl.java
@@ -1,11 +1,11 @@
 package by.st.deliver.core.servicesImpl;
 
 import by.st.deliver.core.dao.RestaurantRepository;
-import by.st.deliver.core.entities.Client;
 import by.st.deliver.core.entities.Restaurant;
-import by.st.deliver.core.mappers.ClientMapper;
 import by.st.deliver.core.mappers.RestaurantMapper;
-import dto.ClientDTO;
+import by.st.deliver.core.servicesImpl.exceptions.DataAlreadyExistException;
+import by.st.deliver.core.servicesImpl.exceptions.NoDataException;
+import by.st.deliver.core.servicesImpl.exceptions.NoSuchDataExceptionQ;
 import dto.RestaurantDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +13,7 @@ import services.RestaurantService;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class RestaurantServiceImpl implements RestaurantService {
@@ -20,41 +21,50 @@ public class RestaurantServiceImpl implements RestaurantService {
     RestaurantRepository restaurantRepository;
 
     @Override
-    public RestaurantDTO addRestaurant(RestaurantDTO restaurantDTO) {
-        if (restaurantRepository.findById(restaurantDTO.getRestaurantId()) == null) {
-            restaurantRepository.save(RestaurantMapper.INSTANCE.restaurantDTOToRestaurant(restaurantDTO));
-            return restaurantDTO;
+    public Long addRestaurant(RestaurantDTO restaurantDTO) {
+        if (restaurantRepository.findById(restaurantDTO.getRestaurantId()) != null) {
+            throw new DataAlreadyExistException("Restaurant with id " + restaurantDTO.getRestaurantId() + "already exists");
         }
-        return null;
+        restaurantRepository.save(RestaurantMapper.INSTANCE.restaurantDTOToRestaurant(restaurantDTO));
+        return restaurantDTO.getRestaurantId();
     }
 
     @Override
     public void removeRestaurant(Long id) {
+
+        if (restaurantRepository.findById(id) == null) {
+            throw new NoSuchDataExceptionQ("There is no restaurant with id " + id);
+        }
         restaurantRepository.deleteById(id);
     }
 
     @Override
     public RestaurantDTO updateRestaurant(RestaurantDTO restaurantDTO) {
         Restaurant restaurant = RestaurantMapper.INSTANCE.restaurantDTOToRestaurant(restaurantDTO);
-        if (restaurantRepository.findById(restaurantDTO.getRestaurantId()) != null) {
-            restaurantRepository.save(restaurant);
-            return restaurantDTO;
-        } else return null;
+        if (restaurantRepository.findById(restaurantDTO.getRestaurantId()) == null) {
+            throw new NoSuchDataExceptionQ("There is no restaurant with id " + restaurantDTO.getRestaurantId());
+        }
+        restaurantRepository.save(restaurant);
+        return restaurantDTO;
     }
 
     @Override
     public List<RestaurantDTO> getRestaurantList() {
         List<Restaurant> restaurants = restaurantRepository.findAll();
-        List<RestaurantDTO> restaurantDTOS = new ArrayList<>();
-        for (int i = 0; i < restaurants.size(); i++) {
-            restaurantDTOS.add(RestaurantMapper.INSTANCE.restaurantToRestaurantDTO(restaurants.get(i)));
+        if (restaurants.isEmpty()) {
+            throw new NoDataException("There are no restaurants on server");
         }
+        List<RestaurantDTO> restaurantDTOS = restaurants.stream().map(restaurant -> RestaurantMapper.INSTANCE.restaurantToRestaurantDTO(restaurant)).collect(Collectors.toList());
         return restaurantDTOS;
     }
 
 
     @Override
     public RestaurantDTO getRestaurantById(Long restaurantId) {
-        return RestaurantMapper.INSTANCE.restaurantToRestaurantDTO(restaurantRepository.findByRestaurantId(restaurantId));
+        RestaurantDTO restaurantDTO = RestaurantMapper.INSTANCE.restaurantToRestaurantDTO(restaurantRepository.findRestaurantByRestaurantId(restaurantId));
+        if (restaurantDTO == null) {
+            throw new NoSuchDataExceptionQ("There is no restaurant with id " + restaurantId);
+        }
+        return restaurantDTO;
     }
 }
